Toolbox: more google-ly navigation
Currently we present the drawer as the primary navigation tool.
But are providing a viewpager as well for top level navigation.
So take it the next step and make the drawer the only top level navigation.

This falls more inline with googles ui revamps (sliding pane
with categories and tabbed viewpagers for subitems)

This allows us to reclaim some screen realestate from the second tabs
and provides swipable navigation for subitems.

We lose subcategories in the drawer but gain more flexablity
overall since we can now include any kind of fragment as a category
and aren't hindered by the limitations of a top level viewpager

Other changes:
  put logcat page into its own category item: bugreport
  grey out settings when toolbox is disabled (excluding
superuser/system)
  fix superuser menu strings
  cleanup imports

Change-Id: I17c4f9d89c05c9db7e0fb3b33ec5b9520496e105