Add the ability for device-specific settings inside the toolbox
Usage example:

  BOARD_DEVICE_SETTINGS := device/lge/hammerhead/DeviceSettings

When a new setting category is needed the class name needs to be added
to DeviceCategory.java in the CLASS_LIST array, along with the string resource
added to strings.xml and the id added to CLASS_NAME_IDS. This will be a simple
list of possible settings fragments, try to keep the naming as general as
possible so we can reuse them.

Make sure to extend DeviceSettingsFragment, this will allow you access
to restore settings on boot completion, and possibly other features down the
road.

Initial imports are DeviceHaptic & DeviceTouchscreen for haptic feedback
and touchscreen tuning.

Change-Id: I80b2af9d06e5c92224b4f51f40e043d7447bb149