Use touch delegate in SwitchBar
The entire switch bar is supposed to behave as a switch,
so it had logic for clicking the layout that duplicated
the logic for the switch itself.

I'm removing all of this duplicate logic and using a
TouchDelegate instead.

This preserves the same behavior with much simpler code.
The previous approach led to accessibility being confused
about exactly what was clickable and what would happen
when different items were clicked. Workarounds to deal with
that confusion created other problems. Sweeping all of it
aside and using a TouchDelegate seems way cleaner.

Bug: 75962891
Test: make SettingsRoboTests
Change-Id: I4fe17d581b5294d2482392f75bf1607126cf235d